+++ pkg/kt/util/init.go
@@ -1,0 +1,6 @@
+package util
+
+func init() {
+	CreateDirIfNotExist(KtHome)
+	FixFileOwner(KtHome)
+}